Move spine buttons to subfolders
