+++ src/NexusMods.UI.Theme/Controls/Spine/Buttons/Game.axaml.cs
@@ -3,15 +3,24 @@ using Avalonia.Controls;
 using Avalonia.Controls.Primitives;
 using Avalonia.Data;
 using Avalonia.Markup.Xaml;
+using Avalonia.Media;
 
-namespace NexusMods.UI.Theme.Controls;
+namespace NexusMods.UI.Theme.Controls.Spine.Buttons;
 
-public partial class SpineAdd : UserControl
+public partial class Game : UserControl
 {
-    public static readonly DirectProperty<SpineAdd, bool?> IsCheckedProperty = AvaloniaProperty.RegisterDirect<SpineAdd, bool?>(nameof(IsChecked),
+    public static readonly StyledProperty<IImage> SourceProperty = AvaloniaProperty.Register<Game, IImage>(nameof(Source));
+    
+    public static readonly DirectProperty<Game, bool?> IsCheckedProperty = AvaloniaProperty.RegisterDirect<Game, bool?>(nameof(IsChecked),
         x => x._toggle?.IsChecked, (x, v) => x._toggle!.IsChecked = v, unsetValue: false, defaultBindingMode: BindingMode.TwoWay);
 
     private readonly ToggleButton? _toggle;
+
+    public IImage Source
+    {
+        get => GetValue(SourceProperty);
+        set => SetValue(SourceProperty, value);
+    }
 
     public bool? IsChecked
     {
@@ -19,8 +28,9 @@ public partial class SpineAdd : UserControl
         set => SetValue(IsCheckedProperty, value);
     }
     
-    public SpineAdd()    {
-        AffectsRender<SpineAdd>(IsCheckedProperty);
+    public Game()
+    {
+        AffectsRender<Home>(IsCheckedProperty, SourceProperty);
         InitializeComponent();
         _toggle = this.Find<ToggleButton>("ToggleButton");
     }
